requirements.txt
torch
torchvision
scikit-learn
streamlit
pillow
numpy
pandas
matplotlib
tqdm
onnx
